# docs/swagger.yaml
definitions:
  base.LogData:
    properties:
      ExecutionTime:
        type: string
      HttpStatusCode:
        type: integer
      IP:
        type: string
      Route:
        type: string
    type: object
  erros.Message:
    properties:
      message:
        type: string
    type: object
  erros.NotAcceptable:
    properties:
      LogData:
        $ref: '#/definitions/base.LogData'
      NotAcceptable:
        $ref: '#/definitions/erros.Message'
    type: object
  inter.ControllerDoctor:
    properties:
      city:
        type: string
      district:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
      street:
        type: string
      zipCode:
        type: string
    type: object
  inter.OutputDoctor:
    properties:
      Login_idLogin:
        type: integer
      city:
        type: string
      create_at:
        type: string
      district:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      street:
        type: string
      update_at:
        type: string
      zipCode:
        type: string
    type: object
  inter.Question:
    properties:
      desc:
        type: string
      number:
        type: string
    type: object
  inter.Resp:
    additionalProperties:
      items:
        $ref: '#/definitions/inter.Question'
      type: array
    type: object
  pInter.Brothers:
    properties:
      health:
        type: string
      neuroDivergent:
        type: string
    type: object
  pInter.Family:
    properties:
      brothers:
        items:
          $ref: '#/definitions/pInter.Brothers'
        type: array
      parents:
        items:
          $ref: '#/definitions/pInter.Parents'
        type: array
    type: object
  pInter.InputPacient:
    properties:
      age:
        type: integer
      email:
        type: string
      family:
        $ref: '#/definitions/pInter.Family'
      gender:
        type: string
      name:
        type: string
      neuroDivergent:
        type: string
      pdfContent:
        items:
          type: integer
        type: array
      phone:
        type: string
      region:
        type: string
      score:
        type: integer
    type: object
  pInter.Pacient:
    properties:
      Doctor_idDoctor:
        type: integer
      Family_idFamily:
        type: integer
      Login_idLogin:
        type: integer
      Skills_idSkills:
        type: integer
      age:
        type: integer
      create_at:
        type: string
      email:
        type: string
      gender:
        type: string
      id:
        type: integer
      name:
        type: string
      neuroDivergent:
        type: string
      pdfContent:
        items:
          type: integer
        type: array
      phone:
        type: string
      region:
        type: string
      score:
        type: integer
      update_at:
        type: string
    type: object
  pInter.Parents:
    properties:
      neuroDivergent:
        type: string
      relationShip:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact: {}
  description: This is a server for app.
  title: Milie Mind
  version: "1.0"
paths:
  /add-pacient:
    post:
      consumes:
      - application/json
      description: Cria um toke para auth do usuario
      parameters:
      - description: Dados do Doutor a ser criado
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pInter.InputPacient'
      - default: Bearer <token>
        description: Token de autenticação (Colocar o token deixando o Bearer)
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'token make:'
          schema:
            $ref: '#/definitions/pInter.Pacient'
      summary: Criar token de auth
      tags:
      - Pacient
  /add-questions:
    post:
      consumes:
      - application/json
      description: Um novo metodo de adicionar as questoes no back
      parameters:
      - description: Dados do Doutor a ser criado
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/inter.Resp'
      produces:
      - application/json
      responses:
        "200":
          description: Doutor criado com sucesso
          schema:
            $ref: '#/definitions/inter.OutputDoctor'
      summary: Adicionar as questoes
      tags:
      - Questions
  /api/create-doctor:
    post:
      consumes:
      - application/json
      description: Cria um novo Doutor no banco de dados
      parameters:
      - description: Dados do Doutor a ser criado
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/inter.ControllerDoctor'
      - default: Bearer <token>
        description: Token de autenticação (Colocar o token deixando o Bearer)
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Doutor criado com sucesso
          schema:
            $ref: '#/definitions/inter.OutputDoctor'
        "406":
          description: Error
          schema:
            $ref: '#/definitions/erros.NotAcceptable'
      summary: Cria um novo Doutor
      tags:
      - Doctor
  /api/modify-doctor:
    put:
      consumes:
      - application/json
      description: Altera o doutor, tem que me passar oque precisa alterar
      parameters:
      - description: Dados do Doutor a ser criado
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/inter.ControllerDoctor'
      - default: Bearer <token>
        description: Token de autenticação (Colocar o token deixando o Bearer)
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Doutor criado com sucesso
          schema:
            $ref: '#/definitions/inter.OutputDoctor'
      summary: Altera Doutor
      tags:
      - Doctor
  /api/pull-questions:
    get:
      responses: {}
  /api/search-doctor:
    get:
      consumes:
      - application/json
      description: Buscar um novo Doutor no banco de dados
      parameters:
      - description: Email do doutor
        in: header
        name: Email
        required: true
        type: string
      - default: Bearer <token>
        description: Token de autenticação (Colocar o token deixando o Bearer)
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Doutor criado com sucesso
          schema:
            $ref: '#/definitions/inter.OutputDoctor'
      summary: Buscar um doutor
      tags:
      - Doctor
  /api/test-token:
    post:
      consumes:
      - application/json
      description: Cria um toke para auth do usuario
      parameters:
      - default: Bearer <token>
        description: Token de autenticação
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'token make:'
          schema:
            type: string
      summary: Test Token
      tags:
      - Auth
  /token:
    get:
      consumes:
      - application/json
      description: Cria um toke para auth do usuario
      produces:
      - application/json
      responses:
        "200":
          description: 'token make:'
          schema:
            type: string
      summary: Criar token de auth
      tags:
      - Auth
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"